First attempt of removing target method from public API
