+++ platforms/software/build-init/src/main/java/org/gradle/api/tasks/wrapper/Wrapper.java
@@ -16,7 +16,6 @@
 
 package org.gradle.api.tasks.wrapper;
 
-import org.gradle.api.DefaultTask;
 import org.gradle.api.GradleException;
 import org.gradle.api.Incubating;
 import org.gradle.api.internal.file.FileLookup;
@@ -29,12 +28,11 @@ import org.gradle.api.tasks.OutputFile;
 import org.gradle.api.tasks.TaskAction;
 import org.gradle.api.tasks.options.Option;
 import org.gradle.api.tasks.options.OptionValues;
-import org.gradle.api.tasks.wrapper.internal.DefaultWrapperVersionsResources;
+import org.gradle.api.tasks.wrapper.internal.InternalWrapper;
 import org.gradle.api.tasks.wrapper.internal.WrapperDefaults;
 import org.gradle.api.tasks.wrapper.internal.WrapperGenerator;
 import org.gradle.internal.UncheckedException;
 import org.gradle.internal.instrumentation.api.annotations.ToBeReplacedByLazyProperty;
-import org.gradle.util.GradleVersion;
 import org.gradle.util.internal.GUtil;
 import org.gradle.util.internal.WrapperDistributionUrlConverter;
 import org.gradle.work.DisableCachingByDefault;
@@ -64,10 +62,10 @@ import java.util.Properties;
  *
  * <p>The scripts generated by this task are intended to be committed to your version control system. This task also
  * generates a small {@code gradle-wrapper.jar} bootstrap JAR file and properties file which should also be committed to
- * your VCS. The scripts delegates to this JAR.
+ * your VCS. The scripts delegate to this JAR.
  */
 @DisableCachingByDefault(because = "Updating the wrapper is not worth caching")
-public abstract class Wrapper extends DefaultTask {
+public abstract class Wrapper extends InternalWrapper {
     public static final String DEFAULT_DISTRIBUTION_PARENT_NAME = WrapperDefaults.DISTRIBUTION_PATH;
 
     /**
@@ -97,7 +95,6 @@ public abstract class Wrapper extends DefaultTask {
     private PathBase distributionBase = WrapperDefaults.DISTRIBUTION_BASE;
     private String distributionUrl;
     private String distributionSha256Sum;
-    private final GradleVersionResolver gradleVersionResolver = new GradleVersionResolver();
     private DistributionType distributionType = WrapperDefaults.DISTRIBUTION_TYPE;
     private String archivePath = WrapperDefaults.ARCHIVE_PATH;
     private PathBase archiveBase = WrapperDefaults.ARCHIVE_BASE;
@@ -182,10 +179,6 @@ public abstract class Wrapper extends DefaultTask {
         } else {
             return null;
         }
-    }
-
-    private boolean isCurrentVersion() {
-        return GradleVersion.current().equals(gradleVersionResolver.getGradleVersion());
     }
 
     /**
@@ -279,20 +272,6 @@ public abstract class Wrapper extends DefaultTask {
     }
 
     /**
-     * Set Wrapper versions resources.
-     *
-     * @since 8.1
-     */
-    @Incubating
-    public void setWrapperVersionsResources(WrapperVersionsResources wrapperVersionsResources) {
-        DefaultWrapperVersionsResources defaultWrapperVersionsResources = (DefaultWrapperVersionsResources) wrapperVersionsResources;
-        gradleVersionResolver.setTextResources(defaultWrapperVersionsResources.getLatest(),
-            defaultWrapperVersionsResources.getReleaseCandidate(),
-            defaultWrapperVersionsResources.getNightly(),
-            defaultWrapperVersionsResources.getReleaseNightly());
-    }
-
-    /**
      * Returns the gradle version for the wrapper.
      *
      * @throws GradleException if the label that can be provided via {@link #setGradleVersion(String)} can not be resolved at the moment. For example, there is not a `release-candidate` available at all times.
@@ -301,7 +280,7 @@ public abstract class Wrapper extends DefaultTask {
     @Input
     @ToBeReplacedByLazyProperty
     public String getGradleVersion() {
-        return gradleVersionResolver.getGradleVersion().getVersion();
+        return getResolvedGradleVersion().getVersion();
     }
 
     /**
@@ -315,7 +294,7 @@ public abstract class Wrapper extends DefaultTask {
         "The following labels are allowed: latest, release-candidate, nightly, and release-nightly.")
     public void setGradleVersion(String gradleVersion) {
         distributionUrlConfigured = true;
-        this.gradleVersionResolver.setGradleVersionString(gradleVersion);
+        setUnresolvedGradleVersion(gradleVersion);
     }
 
     /**
@@ -368,7 +347,7 @@ public abstract class Wrapper extends DefaultTask {
             return distributionUrl;
         }
 
-        return WrapperGenerator.getDistributionUrl(gradleVersionResolver.getGradleVersion(), distributionType);
+        return WrapperGenerator.getDistributionUrl(getResolvedGradleVersion(), distributionType);
     }
 
     /**
