+++ testing/architecture-test/src/changes/archunit-store/public-api-nullability.txt
@@ -193,8 +193,6 @@ Class <org.gradle.api.tasks.testing.logging.TestStackTraceFilter> is not annotat
 Class <org.gradle.api.tasks.testing.testng.TestNGOptions$1> is not annotated (directly or via its package) with @org.jspecify.annotations.NullMarked in (TestNGOptions.java:0)
 Class <org.gradle.api.tasks.testing.testng.TestNGOptions$2> is not annotated (directly or via its package) with @org.jspecify.annotations.NullMarked in (TestNGOptions.java:0)
 Class <org.gradle.api.tasks.testing.testng.TestNGOptions> is not annotated (directly or via its package) with @org.jspecify.annotations.NullMarked in (TestNGOptions.java:0)
-Class <org.gradle.api.tasks.wrapper.GradleVersionResolver$1> is not annotated (directly or via its package) with @org.jspecify.annotations.NullMarked in (GradleVersionResolver.java:0)
-Class <org.gradle.api.tasks.wrapper.GradleVersionResolver> is not annotated (directly or via its package) with @org.jspecify.annotations.NullMarked in (GradleVersionResolver.java:0)
 Class <org.gradle.api.tasks.wrapper.Wrapper$DistributionType> is not annotated (directly or via its package) with @org.jspecify.annotations.NullMarked in (Wrapper.java:0)
 Class <org.gradle.api.tasks.wrapper.Wrapper$PathBase> is not annotated (directly or via its package) with @org.jspecify.annotations.NullMarked in (Wrapper.java:0)
 Class <org.gradle.api.tasks.wrapper.Wrapper> is not annotated (directly or via its package) with @org.jspecify.annotations.NullMarked in (Wrapper.java:0)
